move database entity to database_model package
